[#1329] Refactor reconciler to use BrokerReconciler as canonical type
Refactoring includes the followings tasks:
- Rename ActiveMQArtemisReconcilerImpl to BrokerReconcilerImpl
- Invert the adapter pattern: ActiveMQArtemisReconciler should adapts to BrokerReconciler
- Move BrokerReconcilerImpl functions from activemqartemis_controller.go to broker_reconciler.go
