+++ controllers/activemqartemis_controller_test.go
@@ -3914,8 +3914,8 @@ var _ = Describe("artemis controller", func() {
 			brokerCR, convertErr := ConvertArtemisToBroker(&crd)
 			Expect(convertErr).To(BeNil())
 
-			outer := NewActiveMQArtemisReconciler(k8Manager, ctrl.Log, isOpenshift)
-			reconcilerImpl := NewActiveMQArtemisReconcilerImpl(brokerCR, outer)
+			outer := NewBrokerReconciler(k8Manager, ctrl.Log, isOpenshift)
+			reconcilerImpl := NewBrokerReconcilerImpl(brokerCR, outer)
 
 			defaultConsoleSecretName := crd.Name + "-console-secret"
 			tlsSecret, err := CreateTlsSecret(defaultConsoleSecretName, defaultNamespace, "password", nil)
@@ -4010,8 +4010,8 @@ var _ = Describe("artemis controller", func() {
 			brokerCR, convertErr := ConvertArtemisToBroker(createdCrd)
 			Expect(convertErr).To(BeNil())
 
-			outer := NewActiveMQArtemisReconciler(k8Manager, ctrl.Log, isOpenshift)
-			reconcilerImpl := NewActiveMQArtemisReconcilerImpl(brokerCR, outer)
+			outer := NewBrokerReconciler(k8Manager, ctrl.Log, isOpenshift)
+			reconcilerImpl := NewBrokerReconcilerImpl(brokerCR, outer)
 
 			namers := MakeNamers(brokerCR)
 			defaultConsoleSecretName := createdCrd.Name + "-console-secret"
@@ -4118,8 +4118,8 @@ var _ = Describe("artemis controller", func() {
 			brokerCR, convertErr := ConvertArtemisToBroker(&crd)
 			Expect(convertErr).To(BeNil())
 
-			outer := NewActiveMQArtemisReconciler(k8Manager, ctrl.Log, isOpenshift)
-			reconcilerImpl := NewActiveMQArtemisReconcilerImpl(brokerCR, outer)
+			outer := NewBrokerReconciler(k8Manager, ctrl.Log, isOpenshift)
+			reconcilerImpl := NewBrokerReconcilerImpl(brokerCR, outer)
 			reconcilerImpl.deployed = make(map[reflect.Type][]client.Object)
 
 			namers := MakeNamers(brokerCR)
