+++ controllers/broker_reconciler.go
@@ -31,6 +31,7 @@ import (
 	"github.com/arkmq-org/activemq-artemis-operator/pkg/utils/jolokia_client"
 	"github.com/arkmq-org/activemq-artemis-operator/pkg/utils/namer"
 	"github.com/arkmq-org/activemq-artemis-operator/pkg/utils/random"
+	"github.com/arkmq-org/activemq-artemis-operator/pkg/utils/selectors"
 	"github.com/arkmq-org/activemq-artemis-operator/version"
 	"github.com/go-logr/logr"
 	"github.com/pkg/errors"
@@ -116,7 +117,7 @@ var configCmd = "/opt/amq/bin/launch.sh"
 var defApplyRule string = "merge_all"
 var yacfgProfileVersion = version.YacfgProfileVersionFromFullVersion[version.GetDefaultVersion()]
 
-type ActiveMQArtemisReconcilerImpl struct {
+type BrokerReconcilerImpl struct {
 	requestedResources map[reflect.Type]map[string]rtclient.Object
 	deployed           map[reflect.Type][]rtclient.Object
 	log                logr.Logger
@@ -128,8 +129,8 @@ type ActiveMQArtemisReconcilerImpl struct {
 	matchedTemplates   map[int]bool
 }
 
-func NewActiveMQArtemisReconcilerImpl(customResource *v1beta2.Broker, parent *ActiveMQArtemisReconciler) *ActiveMQArtemisReconcilerImpl {
-	return &ActiveMQArtemisReconcilerImpl{
+func NewBrokerReconcilerImpl(customResource *v1beta2.Broker, parent *BrokerReconciler) *BrokerReconcilerImpl {
+	return &BrokerReconcilerImpl{
 		log:                parent.log,
 		customResource:     customResource,
 		scheme:             parent.Scheme,
@@ -156,7 +157,7 @@ type ActiveMQArtemisIReconciler interface {
 	ProcessResources(customResource *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) uint8
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) Process(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme) error {
+func (reconciler *BrokerReconcilerImpl) Process(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme) error {
 
 	reconciler.log.V(1).Info("Reconciler Processing...", "Operator version", version.Version, "ActiveMQArtemis release", customResource.Spec.Version)
 	reconciler.log.V(2).Info("Reconciler Processing...", "CRD.Name", customResource.Name, "CRD ver", customResource.ObjectMeta.ResourceVersion, "CRD Gen", customResource.ObjectMeta.Generation)
@@ -239,7 +240,7 @@ func trackSecretCheckSumInEnvVar(requestedResources []rtclient.Object, container
 	environments.TrackSecretCheckSumInRollCount(hex.EncodeToString(digest.Sum(nil)), container)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) cloneOfDeployed(kind reflect.Type, name string) rtclient.Object {
+func (reconciler *BrokerReconcilerImpl) cloneOfDeployed(kind reflect.Type, name string) rtclient.Object {
 	obj := reconciler.getFromDeployed(kind, name)
 	if obj != nil {
 		return obj.DeepCopyObject().(rtclient.Object)
@@ -247,7 +248,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) cloneOfDeployed(kind reflect.Ty
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) getFromDeployed(kind reflect.Type, name string) rtclient.Object {
+func (reconciler *BrokerReconcilerImpl) getFromDeployed(kind reflect.Type, name string) rtclient.Object {
 	for _, obj := range reconciler.deployed[kind] {
 		if obj.GetName() == name {
 			return obj
@@ -256,11 +257,11 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) getFromDeployed(kind reflect.Ty
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) addToDeployed(kind reflect.Type, obj rtclient.Object) {
+func (reconciler *BrokerReconcilerImpl) addToDeployed(kind reflect.Type, obj rtclient.Object) {
 	reconciler.deployed[kind] = append(reconciler.deployed[kind], obj)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessStatefulSet(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) (*appsv1.StatefulSet, error) {
+func (reconciler *BrokerReconcilerImpl) ProcessStatefulSet(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) (*appsv1.StatefulSet, error) {
 
 	reqLogger := reconciler.log.WithName(customResource.Name)
 
@@ -322,7 +323,7 @@ func isClustered(customResource *v1beta2.Broker) bool {
 	return true
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessCredentials(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) {
+func (reconciler *BrokerReconcilerImpl) ProcessCredentials(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) {
 
 	if common.IsRestricted(customResource) {
 		return
@@ -377,7 +378,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessCredentials(customResour
 	reconciler.sourceEnvVarFromSecret(customResource, namer, currentStatefulSet, &envVars, secretName, client)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessDeploymentPlan(customResource *v1beta2.Broker, theNamer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) {
+func (reconciler *BrokerReconcilerImpl) ProcessDeploymentPlan(customResource *v1beta2.Broker, theNamer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) {
 
 	deploymentPlan := &customResource.Spec.DeploymentPlan
 
@@ -404,7 +405,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessDeploymentPlan(customRes
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) controllerManagedScaledownViaConditions(customResource *v1beta2.Broker, currentStatefulSet *appsv1.StatefulSet, reqestedReplicas int32, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) controllerManagedScaledownViaConditions(customResource *v1beta2.Broker, currentStatefulSet *appsv1.StatefulSet, reqestedReplicas int32, client rtclient.Client) {
 	if isClustered(customResource) &&
 		*customResource.Spec.DeploymentPlan.MessageMigration &&
 		currentStatefulSet != nil &&
@@ -476,11 +477,11 @@ func OrdinalToDrain(currentStatefulSet *appsv1.StatefulSet) int32 {
 	return *currentStatefulSet.Spec.Replicas - 1
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) CrConfiguredForControllerManagedScaleDown() bool {
+func (reconciler *BrokerReconcilerImpl) CrConfiguredForControllerManagedScaleDown() bool {
 	return slices.Contains(reconciler.customResource.Spec.BrokerProperties, ScaleDownConfigTrigger)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) shutdownWithScaledown(client rtclient.Client, ordinalToDrain int32) error {
+func (reconciler *BrokerReconcilerImpl) shutdownWithScaledown(client rtclient.Client, ordinalToDrain int32) error {
 
 	if err := reconciler.resolveAndValidateJolokia(client, ordinalToDrain); err != nil {
 		return err
@@ -495,7 +496,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) shutdownWithScaledown(client rt
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) GetTotalMessageCount(customResource *v1beta2.Broker, client rtclient.Client, ordinalToDrain int32) (int64, error) {
+func (reconciler *BrokerReconcilerImpl) GetTotalMessageCount(customResource *v1beta2.Broker, client rtclient.Client, ordinalToDrain int32) (int64, error) {
 	if err := reconciler.resolveAndValidateJolokia(client, ordinalToDrain); err != nil {
 		return -1, err
 	}
@@ -508,7 +509,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) GetTotalMessageCount(customReso
 	return strconv.ParseInt(result, 10, 64)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) resolveAndValidateJolokia(client rtclient.Client, ordinalOfInterest int32) error {
+func (reconciler *BrokerReconcilerImpl) resolveAndValidateJolokia(client rtclient.Client, ordinalOfInterest int32) error {
 	reconciler.resolveJolokiaEndpoints(reconciler.customResource, client)
 	if len(reconciler.jolokiaEndpoints) == 0 {
 		return NewJolokiaClientsNotFoundError(errors.New("waiting for jolokia endpoints to become available"))
@@ -519,7 +520,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) resolveAndValidateJolokia(clien
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) applyPodDisruptionBudget(customResource *v1beta2.Broker) {
+func (reconciler *BrokerReconcilerImpl) applyPodDisruptionBudget(customResource *v1beta2.Broker) {
 
 	var desired *policyv1.PodDisruptionBudget
 	obj := reconciler.cloneOfDeployed(reflect.TypeOf(policyv1.PodDisruptionBudget{}), customResource.Name+"-pdb")
@@ -548,7 +549,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) applyPodDisruptionBudget(custom
 	reconciler.trackDesired(desired)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessAcceptorsAndConnectors(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) error {
+func (reconciler *BrokerReconcilerImpl) ProcessAcceptorsAndConnectors(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) error {
 
 	if common.IsRestricted(customResource) {
 		return nil
@@ -583,7 +584,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessAcceptorsAndConnectors(c
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessConsole(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) error {
+func (reconciler *BrokerReconcilerImpl) ProcessConsole(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme, currentStatefulSet *appsv1.StatefulSet) error {
 
 	reconciler.configureConsoleExposure(customResource, namer, client)
 	if !customResource.Spec.Console.SSLEnabled || common.IsRestricted(customResource) {
@@ -622,7 +623,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessConsole(customResource *
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) appendSystemPropertiesForConsole(currentSS *appsv1.StatefulSet, systemArgs string) {
+func (reconciler *BrokerReconcilerImpl) appendSystemPropertiesForConsole(currentSS *appsv1.StatefulSet, systemArgs string) {
 	reconciler.log.V(1).Info("Appending console system prop", "value", systemArgs)
 	consoleProps := corev1.EnvVar{
 		Name:  javaArgsAppendEnvVarName,
@@ -631,7 +632,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) appendSystemPropertiesForConsol
 	environments.CreateOrAppend(currentSS.Spec.Template.Spec.Containers, &consoleProps)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) syncMessageMigration(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme) {
+func (reconciler *BrokerReconcilerImpl) syncMessageMigration(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client, scheme *runtime.Scheme) {
 
 	var err error = nil
 	var retrieveError error = nil
@@ -704,7 +705,7 @@ func isLocalOnly() bool {
 	return oprNamespace == watchNamespace
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) sourceEnvVarFromSecret(customResource *v1beta2.Broker, namer common.Namers, currentStatefulSet *appsv1.StatefulSet, envVars *map[string]ValueInfo, secretName string, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) sourceEnvVarFromSecret(customResource *v1beta2.Broker, namer common.Namers, currentStatefulSet *appsv1.StatefulSet, envVars *map[string]ValueInfo, secretName string, client rtclient.Client) {
 
 	var log = reconciler.log.WithName("controller_v1beta1activemqartemis").WithName("sourceEnvVarFromSecret")
 
@@ -841,7 +842,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) sourceEnvVarFromSecret(customRe
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) processSSLSecret(secretName string, customResource *v1beta2.Broker, client rtclient.Client) (*corev1.Secret, error) {
+func (reconciler *BrokerReconcilerImpl) processSSLSecret(secretName string, customResource *v1beta2.Broker, client rtclient.Client) (*corev1.Secret, error) {
 
 	if strings.HasSuffix(secretName, certutil.Cert_provided_secret_suffix) {
 		return &corev1.Secret{ObjectMeta: metav1.ObjectMeta{Name: secretName}}, nil
@@ -871,7 +872,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) processSSLSecret(secretName str
 	return trackedSecret, nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) generateAcceptorsString(customResource *v1beta2.Broker, client rtclient.Client, currentSS *appsv1.StatefulSet) (string, error) {
+func (reconciler *BrokerReconcilerImpl) generateAcceptorsString(customResource *v1beta2.Broker, client rtclient.Client, currentSS *appsv1.StatefulSet) (string, error) {
 
 	// TODO: Optimize for the single broker configuration
 	ensureCOREOn61616Exists := true // as clustered is no longer an option but true by default
@@ -980,7 +981,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) generateAcceptorsString(customR
 	return acceptorEntry, nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) generateConnectorsString(customResource *v1beta2.Broker, client rtclient.Client, currentSS *appsv1.StatefulSet) (string, error) {
+func (reconciler *BrokerReconcilerImpl) generateConnectorsString(customResource *v1beta2.Broker, client rtclient.Client, currentSS *appsv1.StatefulSet) (string, error) {
 
 	connectorEntry := ""
 	connectors := customResource.Spec.Connectors
@@ -1025,7 +1026,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) generateConnectorsString(custom
 	return connectorEntry, nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureAcceptorsExposure(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) configureAcceptorsExposure(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) {
 	originalLabels := namer.LabelBuilder.Labels()
 	namespacedName := types.NamespacedName{
 		Name:      customResource.Name,
@@ -1055,7 +1056,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) configureAcceptorsExposure(cust
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ServiceDefinitionForCR(serviceName types.NamespacedName, client rtclient.Client, nameSuffix string, portNumber int32, selectorLabels map[string]string, labels map[string]string) *corev1.Service {
+func (reconciler *BrokerReconcilerImpl) ServiceDefinitionForCR(serviceName types.NamespacedName, client rtclient.Client, nameSuffix string, portNumber int32, selectorLabels map[string]string, labels map[string]string) *corev1.Service {
 	var serviceDefinition *corev1.Service
 	obj := reconciler.cloneOfDeployed(reflect.TypeOf(corev1.Service{}), serviceName.Name)
 	if obj != nil {
@@ -1064,7 +1065,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ServiceDefinitionForCR(serviceN
 	return svc.NewServiceDefinitionForCR(serviceName, client, nameSuffix, portNumber, selectorLabels, labels, serviceDefinition)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ExposureDefinitionForCR(customResource *v1beta2.Broker, namespacedName types.NamespacedName, labels map[string]string, passthroughTLS bool, ingressHost string, ordinalString string, itemName string, exposeMode *v1beta2.ExposeMode) rtclient.Object {
+func (reconciler *BrokerReconcilerImpl) ExposureDefinitionForCR(customResource *v1beta2.Broker, namespacedName types.NamespacedName, labels map[string]string, passthroughTLS bool, ingressHost string, ordinalString string, itemName string, exposeMode *v1beta2.ExposeMode) rtclient.Object {
 
 	targetPortName := itemName + "-" + ordinalString
 	targetServiceName := customResource.Name + "-" + targetPortName + "-" + ServiceTypePostfix
@@ -1094,7 +1095,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ExposureDefinitionForCR(customR
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) trackDesired(desired rtclient.Object) {
+func (reconciler *BrokerReconcilerImpl) trackDesired(desired rtclient.Object) {
 	desiredType := reflect.TypeOf(desired)
 	if reconciler.requestedResources == nil {
 		reconciler.requestedResources = make(map[reflect.Type]map[string]rtclient.Object)
@@ -1108,7 +1109,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) trackDesired(desired rtclient.O
 	resMap[resName] = desired
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) getFromDesired(kind reflect.Type, name string) rtclient.Object {
+func (reconciler *BrokerReconcilerImpl) getFromDesired(kind reflect.Type, name string) rtclient.Object {
 	obj, found := reconciler.requestedResources[kind][name]
 	if found {
 		return obj
@@ -1116,7 +1117,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) getFromDesired(kind reflect.Typ
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) applyTemplates(desired rtclient.Object) (err error) {
+func (reconciler *BrokerReconcilerImpl) applyTemplates(desired rtclient.Object) (err error) {
 	for index, template := range reconciler.customResource.Spec.ResourceTemplates {
 		if err = reconciler.applyTemplate(index, template, desired); err != nil {
 			break
@@ -1125,7 +1126,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) applyTemplates(desired rtclient
 	return err
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) applyTemplate(index int, template v1beta2.ResourceTemplate, target rtclient.Object) error {
+func (reconciler *BrokerReconcilerImpl) applyTemplate(index int, template v1beta2.ResourceTemplate, target rtclient.Object) error {
 	if match(template, target) {
 
 		reconciler.matchedTemplates[index] = true
@@ -1199,7 +1200,7 @@ func formatTemplatedObject(customResource *v1beta2.Broker, object interface{}, o
 	return object
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) applyFormattedKeyValue(collection map[string]string, ordinal string, itemName string, resType string, key string, value string) {
+func (reconciler *BrokerReconcilerImpl) applyFormattedKeyValue(collection map[string]string, ordinal string, itemName string, resType string, key string, value string) {
 	formattedKey := formatTemplatedString(reconciler.customResource, key, ordinal, itemName, resType)
 	if value == RemoveKeySpecialValue {
 		delete(collection, formattedKey)
@@ -1305,7 +1306,7 @@ func match(template v1beta2.ResourceTemplate, target rtclient.Object) bool {
 	return true
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureConnectorsExposure(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) configureConnectorsExposure(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) {
 	originalLabels := namer.LabelBuilder.Labels()
 	namespacedName := types.NamespacedName{
 		Name:      customResource.Name,
@@ -1338,7 +1339,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) configureConnectorsExposure(cus
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureConsoleExposure(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) configureConsoleExposure(customResource *v1beta2.Broker, namer common.Namers, client rtclient.Client) {
 	console := customResource.Spec.Console
 	consoleName := customResource.Spec.Console.Name
 
@@ -1417,7 +1418,7 @@ func formatTemplatedString(customResource *v1beta2.Broker, template string, brok
 	return template
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) generateCommonSSLFlags(customResource *v1beta2.Broker, secretName string, caSecretName *string, trustStoreType string, client rtclient.Client, isConsole bool) (*certutil.SslArguments, string, error) {
+func (reconciler *BrokerReconcilerImpl) generateCommonSSLFlags(customResource *v1beta2.Broker, secretName string, caSecretName *string, trustStoreType string, client rtclient.Client, isConsole bool) (*certutil.SslArguments, string, error) {
 
 	sslSecret := &corev1.Secret{}
 	if strings.HasSuffix(secretName, certutil.Cert_provided_secret_suffix) {
@@ -1459,7 +1460,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) generateCommonSSLFlags(customRe
 	return sslArgs, sslFlags, nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) generateAcceptorSSLOptionalArguments(acceptor v1beta2.AcceptorType) string {
+func (reconciler *BrokerReconcilerImpl) generateAcceptorSSLOptionalArguments(acceptor v1beta2.AcceptorType) string {
 
 	sslOptionalArguments := ""
 
@@ -1496,7 +1497,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) generateAcceptorSSLOptionalArgu
 	return sslOptionalArguments
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) addPemConfigFileSecret(ss *appsv1.StatefulSet, configs []string) {
+func (reconciler *BrokerReconcilerImpl) addPemConfigFileSecret(ss *appsv1.StatefulSet, configs []string) {
 	resourceName := types.NamespacedName{
 		Namespace: ss.Namespace,
 		Name:      certutil.CfgToSecretName(configs[0]),
@@ -1530,7 +1531,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) addPemConfigFileSecret(ss *apps
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) generateConnectorSSLOptionalArguments(connector v1beta2.ConnectorType) string {
+func (reconciler *BrokerReconcilerImpl) generateConnectorSSLOptionalArguments(connector v1beta2.ConnectorType) string {
 
 	sslOptionalArguments := ""
 
@@ -1571,7 +1572,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) generateConnectorSSLOptionalArg
 	return sslOptionalArguments
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) CurrentDeployedResources(customResource *v1beta2.Broker, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) CurrentDeployedResources(customResource *v1beta2.Broker, client rtclient.Client) {
 	reqLogger := reconciler.log.WithValues("ActiveMQArtemis Name", customResource.Name)
 
 	var err error
@@ -1600,7 +1601,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) CurrentDeployedResources(custom
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessResources(customResource *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) (err error) {
+func (reconciler *BrokerReconcilerImpl) ProcessResources(customResource *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) (err error) {
 
 	reqLogger := reconciler.log.WithValues("ActiveMQArtemis Name", customResource.Name)
 
@@ -1679,21 +1680,21 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessResources(customResource
 	}
 }
 
-func countOfRequested(reconciler *ActiveMQArtemisReconcilerImpl) (total int) {
+func countOfRequested(reconciler *BrokerReconcilerImpl) (total int) {
 	for _, v := range reconciler.requestedResources {
 		total += len(v)
 	}
 	return total
 }
 
-func countOfDeployed(reconciler *ActiveMQArtemisReconcilerImpl) (total int) {
+func countOfDeployed(reconciler *BrokerReconcilerImpl) (total int) {
 	for _, v := range reconciler.deployed {
 		total += len(v)
 	}
 	return total
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) CompareMetaAndSpec(deployed, requested rtclient.Object) bool {
+func (reconciler *BrokerReconcilerImpl) CompareMetaAndSpec(deployed, requested rtclient.Object) bool {
 
 	isEqual := equalObjectMeta(deployed, requested) &&
 		equality.Semantic.DeepEqual(specOf(deployed), specOf(requested)) &&
@@ -1704,7 +1705,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) CompareMetaAndSpec(deployed, re
 	return isEqual
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) CompareSecret(deployed, requested rtclient.Object) bool {
+func (reconciler *BrokerReconcilerImpl) CompareSecret(deployed, requested rtclient.Object) bool {
 
 	isEqual := equalObjectMeta(deployed, requested) &&
 		reconciler.ensureOwnerReferenceAPIVersion(reconciler.customResource, deployed, requested)
@@ -1725,7 +1726,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) CompareSecret(deployed, request
 	return isEqual
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) CompareConfigMap(deployed, requested rtclient.Object) bool {
+func (reconciler *BrokerReconcilerImpl) CompareConfigMap(deployed, requested rtclient.Object) bool {
 	// our single configMap is immutable, the name indicates a change
 	return deployed.GetName() == requested.GetName() &&
 		reconciler.ensureOwnerReferenceAPIVersion(reconciler.customResource, deployed, requested)
@@ -1770,28 +1771,28 @@ func trackError(compositeError *[]error, err error) {
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) createResource(customResource *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme, requested rtclient.Object, kind reflect.Type) error {
+func (reconciler *BrokerReconcilerImpl) createResource(customResource *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme, requested rtclient.Object, kind reflect.Type) error {
 	reconciler.log.V(1).Info("Adding delta resources, i.e. creating ", "name ", requested.GetName(), "of kind ", kind)
 	return reconciler.createRequestedResource(customResource, client, scheme, requested, kind)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) updateResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
+func (reconciler *BrokerReconcilerImpl) updateResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
 	reconciler.log.V(1).Info("Updating delta resources, i.e. updating ", "name ", requested.GetName(), "of kind ", kind)
 	return reconciler.updateRequestedResource(client, requested, kind)
 
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) deleteResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
+func (reconciler *BrokerReconcilerImpl) deleteResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
 	reconciler.log.V(1).Info("Deleting delta resources, i.e. removing ", "name ", requested.GetName(), "of kind ", kind)
 	return reconciler.deleteRequestedResource(client, requested, kind)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) createRequestedResource(customResource *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme, requested rtclient.Object, kind reflect.Type) error {
+func (reconciler *BrokerReconcilerImpl) createRequestedResource(customResource *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme, requested rtclient.Object, kind reflect.Type) error {
 	reconciler.log.V(1).Info("Creating ", "kind ", kind, "named ", requested.GetName())
 	return resources.Create(customResource, client, scheme, requested)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) updateRequestedResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
+func (reconciler *BrokerReconcilerImpl) updateRequestedResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
 	var updateError error
 	if updateError = resources.Update(client, requested); updateError == nil {
 		reconciler.log.V(1).Info("updated", "kind ", kind, "named ", requested.GetName())
@@ -1801,7 +1802,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) updateRequestedResource(client 
 	return updateError
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) deleteRequestedResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
+func (reconciler *BrokerReconcilerImpl) deleteRequestedResource(client rtclient.Client, requested rtclient.Object, kind reflect.Type) error {
 
 	var deleteError error
 	if deleteError := resources.Delete(client, requested); deleteError == nil {
@@ -1813,7 +1814,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) deleteRequestedResource(client 
 }
 
 // older version of the operator would drop the owner reference, we need to adopt such secrets and update them
-func (reconciler *ActiveMQArtemisReconcilerImpl) adoptExistingSecretWithNoOwnerRefForUpdate(cr *v1beta2.Broker, candidate *corev1.Secret, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) adoptExistingSecretWithNoOwnerRefForUpdate(cr *v1beta2.Broker, candidate *corev1.Secret, client rtclient.Client) {
 
 	key := types.NamespacedName{Name: candidate.Name, Namespace: candidate.Namespace}
 	existingSecret := &corev1.Secret{}
@@ -1827,7 +1828,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) adoptExistingSecretWithNoOwnerR
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) updateOwnerReferencesAndMatchVersion(cr *v1beta2.Broker, existing rtclient.Object, candidate rtclient.Object) {
+func (reconciler *BrokerReconcilerImpl) updateOwnerReferencesAndMatchVersion(cr *v1beta2.Broker, existing rtclient.Object, candidate rtclient.Object) {
 
 	resources.SetOwnerAndController(cr, existing)
 	candidate.SetOwnerReferences(existing.GetOwnerReferences())
@@ -1836,7 +1837,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) updateOwnerReferencesAndMatchVe
 	candidate.SetUID(existing.GetUID())
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ensureOwnerReferenceAPIVersion(cr *v1beta2.Broker, existing rtclient.Object, candidate rtclient.Object) bool {
+func (reconciler *BrokerReconcilerImpl) ensureOwnerReferenceAPIVersion(cr *v1beta2.Broker, existing rtclient.Object, candidate rtclient.Object) bool {
 	ownerRefs := existing.GetOwnerReferences()
 	if len(ownerRefs) > 0 {
 		for i := range ownerRefs {
@@ -1856,7 +1857,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ensureOwnerReferenceAPIVersion(
 	return true
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) checkExistingService(cr *v1beta2.Broker, candidate *corev1.Service, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) checkExistingService(cr *v1beta2.Broker, candidate *corev1.Service, client rtclient.Client) {
 	serviceType := reflect.TypeOf(corev1.Service{})
 	obj := reconciler.getFromDeployed(serviceType, candidate.Name)
 	if obj != nil {
@@ -1878,7 +1879,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) checkExistingService(cr *v1beta
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) checkExistingPersistentVolumes(instance *v1beta2.Broker, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) checkExistingPersistentVolumes(instance *v1beta2.Broker, client rtclient.Client) {
 	var i int32
 	for i = 0; i < common.GetDeploymentSize(instance); i++ {
 		ordinalString := strconv.Itoa(int(i))
@@ -1940,7 +1941,7 @@ func addNewVolumes(existingNames map[string]string, existing *[]corev1.Volume, n
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) MakeVolumes(customResource *v1beta2.Broker, namer common.Namers) ([]corev1.Volume, error) {
+func (reconciler *BrokerReconcilerImpl) MakeVolumes(customResource *v1beta2.Broker, namer common.Namers) ([]corev1.Volume, error) {
 
 	volumeDefinitions := []corev1.Volume{}
 	if customResource.Spec.DeploymentPlan.PersistenceEnabled {
@@ -2054,7 +2055,7 @@ func MakeExtraVolumeMounts(customResource *v1beta2.Broker) []corev1.VolumeMount 
 	return volumeMounts
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) MakeVolumeMounts(customResource *v1beta2.Broker, namer common.Namers) ([]corev1.VolumeMount, error) {
+func (reconciler *BrokerReconcilerImpl) MakeVolumeMounts(customResource *v1beta2.Broker, namer common.Namers) ([]corev1.VolumeMount, error) {
 
 	volumeMounts := []corev1.VolumeMount{}
 	if customResource.Spec.DeploymentPlan.PersistenceEnabled || common.IsRestricted(customResource) {
@@ -2139,7 +2140,7 @@ func MakeContainerPorts(cr *v1beta2.Broker) []corev1.ContainerPort {
 	return containerPorts
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) PodTemplateSpecForCR(customResource *v1beta2.Broker, namer common.Namers, currentStatefulSet *appsv1.StatefulSet, client rtclient.Client) (*corev1.PodTemplateSpec, error) {
+func (reconciler *BrokerReconcilerImpl) PodTemplateSpecForCR(customResource *v1beta2.Broker, namer common.Namers, currentStatefulSet *appsv1.StatefulSet, client rtclient.Client) (*corev1.PodTemplateSpec, error) {
 
 	reqLogger := reconciler.log.WithName(customResource.Name)
 
@@ -2799,7 +2800,7 @@ func newBufferWithHeader(commentChars string) *bytes.Buffer {
 	return buf
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) brokerPropertiesConfigSystemPropValue(mountPoint, resourceName string, brokerPropertiesData map[string][]byte) string {
+func (reconciler *BrokerReconcilerImpl) brokerPropertiesConfigSystemPropValue(mountPoint, resourceName string, brokerPropertiesData map[string][]byte) string {
 	var result = ""
 	if len(brokerPropertiesData) == 1 && !reconciler.CrConfiguredForControllerManagedScaleDown() {
 		// single entry, no ordinal subpath - broker will log if arg is not found for the watcher so make conditional
@@ -2852,7 +2853,7 @@ func getConfigExtraMount(customResource *v1beta2.Broker, suffix string) (string,
 	return "", "", false
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureStartupProbe(container *corev1.Container, probeFromCr *corev1.Probe) *corev1.Probe {
+func (reconciler *BrokerReconcilerImpl) configureStartupProbe(container *corev1.Container, probeFromCr *corev1.Probe) *corev1.Probe {
 
 	var startupProbe *corev1.Probe = container.StartupProbe
 	reconciler.log.V(1).Info("Configuring Startup Probe", "existing", startupProbe)
@@ -2871,7 +2872,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) configureStartupProbe(container
 	return startupProbe
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureLivenessProbe(container *corev1.Container, probeFromCr *corev1.Probe) *corev1.Probe {
+func (reconciler *BrokerReconcilerImpl) configureLivenessProbe(container *corev1.Container, probeFromCr *corev1.Probe) *corev1.Probe {
 	var livenessProbe *corev1.Probe = container.LivenessProbe
 	reconciler.log.V(1).Info("Configuring Liveness Probe", "existing", livenessProbe)
 
@@ -2923,7 +2924,7 @@ var betterCommand = []string{
 	// "1", sleep seconds not applicable with 1 retry
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureReadinessProbe(container *corev1.Container, probeFromCr *corev1.Probe) *corev1.Probe {
+func (reconciler *BrokerReconcilerImpl) configureReadinessProbe(container *corev1.Container, probeFromCr *corev1.Probe) *corev1.Probe {
 
 	var readinessProbe *corev1.Probe = container.ReadinessProbe
 	reconciler.log.V(1).Info("Configuring Readyness Probe", "existing", readinessProbe)
@@ -3046,7 +3047,7 @@ func getPropertiesResourceNsName(artemis *v1beta2.Broker) types.NamespacedName {
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) addResourceForBrokerProperties(customResource *v1beta2.Broker, namer common.Namers) (string, bool, map[string][]byte, error) {
+func (reconciler *BrokerReconcilerImpl) addResourceForBrokerProperties(customResource *v1beta2.Broker, namer common.Namers) (string, bool, map[string][]byte, error) {
 
 	// fetch and do idempotent transform based on CR
 
@@ -3094,7 +3095,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) addResourceForBrokerProperties(
 	return resourceName.Name, true, data, nil
 }
 
-func (r *ActiveMQArtemisReconcilerImpl) ProcessBrokerProperties(m map[string][]byte) {
+func (r *BrokerReconcilerImpl) ProcessBrokerProperties(m map[string][]byte) {
 	if condition := meta.FindStatusCondition(r.customResource.Status.Conditions, v1beta2.ScaleDownPendingConditionType); condition != nil {
 		if ordinal, err := r.ordinalFromScaleDownCondition(condition); err == nil {
 			buf := NewPropsWithHeader()
@@ -3156,7 +3157,7 @@ func BrokerPropertiesData(props []string) map[string][]byte {
 	return contents
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureAffinity(podSpec *corev1.PodSpec, affinity *v1beta2.AffinityConfig) {
+func (reconciler *BrokerReconcilerImpl) configureAffinity(podSpec *corev1.PodSpec, affinity *v1beta2.AffinityConfig) {
 	if affinity != nil {
 		podSpec.Affinity = &corev1.Affinity{}
 		if affinity.PodAffinity != nil {
@@ -3174,7 +3175,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) configureAffinity(podSpec *core
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configurePodSecurityContext(podSpec *corev1.PodSpec, podSecurityContext *corev1.PodSecurityContext) {
+func (reconciler *BrokerReconcilerImpl) configurePodSecurityContext(podSpec *corev1.PodSpec, podSecurityContext *corev1.PodSecurityContext) {
 	reconciler.log.V(1).Info("Configuring PodSecurityContext")
 
 	if nil != podSecurityContext {
@@ -3191,7 +3192,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) configurePodSecurityContext(pod
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configureContianerSecurityContext(container *corev1.Container, containerSecurityContext *corev1.SecurityContext) {
+func (reconciler *BrokerReconcilerImpl) configureContianerSecurityContext(container *corev1.Container, containerSecurityContext *corev1.SecurityContext) {
 	reconciler.log.V(1).Info("Configuring Container SecurityContext")
 
 	if nil != containerSecurityContext {
@@ -3234,7 +3235,7 @@ func sortedKeysStringKeyByteValue(props map[string][]byte) []string {
 	return sortedKeys
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) configPodSecurity(podSpec *corev1.PodSpec, podSecurity *v1beta2.PodSecurityType) {
+func (reconciler *BrokerReconcilerImpl) configPodSecurity(podSpec *corev1.PodSpec, podSecurity *v1beta2.PodSecurityType) {
 	if podSecurity.ServiceAccountName != nil {
 		reconciler.log.V(2).Info("Pod serviceAccountName specified", "existing", podSpec.ServiceAccountName, "new", *podSecurity.ServiceAccountName)
 		podSpec.ServiceAccountName = *podSecurity.ServiceAccountName
@@ -3259,7 +3260,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) configPodSecurity(podSpec *core
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) createExtraConfigmapsAndSecretsVolumeMounts(configMaps []string, secrets []string, brokePropertiesResourceName string, brokerPropsData map[string][]byte, client rtclient.Client) ([]corev1.Volume, []corev1.VolumeMount, error) {
+func (reconciler *BrokerReconcilerImpl) createExtraConfigmapsAndSecretsVolumeMounts(configMaps []string, secrets []string, brokePropertiesResourceName string, brokerPropsData map[string][]byte, client rtclient.Client) ([]corev1.Volume, []corev1.VolumeMount, error) {
 
 	var extraVolumes []corev1.Volume
 	var extraVolumeMounts []corev1.VolumeMount
@@ -3354,7 +3355,7 @@ func ParseBrokerPropertyWithOrdinal(property string) []string {
 	return brokerPropertyWithOrdinalRegex.FindStringSubmatch(property)
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) StatefulSetForCR(customResource *v1beta2.Broker, namer common.Namers, currentStateFullSet *appsv1.StatefulSet, client rtclient.Client) (*appsv1.StatefulSet, error) {
+func (reconciler *BrokerReconcilerImpl) StatefulSetForCR(customResource *v1beta2.Broker, namer common.Namers, currentStateFullSet *appsv1.StatefulSet, client rtclient.Client) (*appsv1.StatefulSet, error) {
 
 	//	reqLogger := reconciler.log.WithName(customResource.Name)
 
@@ -3377,7 +3378,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) StatefulSetForCR(customResource
 	return currentStateFullSet, nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) PersistentVolumeClaimArrayForCR(customResource *v1beta2.Broker, namer common.Namers, spec appsv1.StatefulSetSpec) []corev1.PersistentVolumeClaim {
+func (reconciler *BrokerReconcilerImpl) PersistentVolumeClaimArrayForCR(customResource *v1beta2.Broker, namer common.Namers, spec appsv1.StatefulSetSpec) []corev1.PersistentVolumeClaim {
 
 	var existing, current *corev1.PersistentVolumeClaim
 	pvcArray := make([]corev1.PersistentVolumeClaim, 0)
@@ -3530,7 +3531,7 @@ type applyError struct {
 	Reason       string `json:"reason"`
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessBrokerStatus(cr *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) (retry bool) {
+func (reconciler *BrokerReconcilerImpl) ProcessBrokerStatus(cr *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) (retry bool) {
 	var condition metav1.Condition
 
 	// we need to requeue till stable
@@ -3624,11 +3625,11 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) ProcessBrokerStatus(cr *v1beta2
 	return retry
 }
 
-func (r *ActiveMQArtemisReconcilerImpl) ScaleDownConditionPodNameMessage(ordinal int32) string {
+func (r *BrokerReconcilerImpl) ScaleDownConditionPodNameMessage(ordinal int32) string {
 	return r.ScaleDownConditionMessage(ordinal, "")
 }
 
-func (r *ActiveMQArtemisReconcilerImpl) ScaleDownConditionMessage(ordinal int32, message string) string {
+func (r *BrokerReconcilerImpl) ScaleDownConditionMessage(ordinal int32, message string) string {
 	podName := namer.CrToSSOrdinal(r.customResource.Name, int(ordinal))
 	if message == "" {
 		return podName
@@ -3637,7 +3638,7 @@ func (r *ActiveMQArtemisReconcilerImpl) ScaleDownConditionMessage(ordinal int32,
 	}
 }
 
-func (r *ActiveMQArtemisReconcilerImpl) ordinalFromScaleDownCondition(condition *metav1.Condition) (int, error) {
+func (r *BrokerReconcilerImpl) ordinalFromScaleDownCondition(condition *metav1.Condition) (int, error) {
 	return namer.OrdinalFromPodName(r.customResource.Name, condition.Message)
 }
 
@@ -3715,7 +3716,7 @@ func AssertBrokersAvailable(cr *v1beta2.Broker, client rtclient.Client) ArtemisE
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) AssertBrokerPropertiesStatus(cr *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) ArtemisError {
+func (reconciler *BrokerReconcilerImpl) AssertBrokerPropertiesStatus(cr *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) ArtemisError {
 	reqLogger := ctrl.Log.WithValues("ActiveMQArtemis Name", cr.Name)
 
 	secretProjection, err := reconciler.getSecretProjection(getPropertiesResourceNsName(cr), client)
@@ -3754,7 +3755,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) AssertBrokerPropertiesStatus(cr
 	return errorStatus
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) AssertJaasPropertiesStatus(cr *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) ArtemisError {
+func (reconciler *BrokerReconcilerImpl) AssertJaasPropertiesStatus(cr *v1beta2.Broker, client rtclient.Client, scheme *runtime.Scheme) ArtemisError {
 	reqLogger := ctrl.Log.WithValues("ActiveMQArtemis Name", cr.Name)
 
 	Projection, err := reconciler.getConfigMappedJaasProperties(cr, client)
@@ -3775,7 +3776,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) AssertJaasPropertiesStatus(cr *
 	return statusError
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) AssertBrokerImageVersion(cr *v1beta2.Broker, client rtclient.Client) ArtemisError {
+func (reconciler *BrokerReconcilerImpl) AssertBrokerImageVersion(cr *v1beta2.Broker, client rtclient.Client) ArtemisError {
 	reqLogger := ctrl.Log.WithValues("ActiveMQArtemis Name", cr.Name)
 
 	// The ResolveBrokerVersionFromCR should never fail because validation succeeded
@@ -3796,7 +3797,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) AssertBrokerImageVersion(cr *v1
 	return statusError
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) CheckStatus(cr *v1beta2.Broker, client rtclient.Client, checkBrokerStatus func(BrokerStatus *brokerStatus, jk *jolokia_client.JkInfo) ArtemisError) ArtemisError {
+func (reconciler *BrokerReconcilerImpl) CheckStatus(cr *v1beta2.Broker, client rtclient.Client, checkBrokerStatus func(BrokerStatus *brokerStatus, jk *jolokia_client.JkInfo) ArtemisError) ArtemisError {
 
 	reconciler.resolveJolokiaEndpoints(cr, client)
 
@@ -3816,7 +3817,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) CheckStatus(cr *v1beta2.Broker,
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) CheckStatusFromJolokia(jk *jolokia_client.JkInfo, checkBrokerStatus func(BrokerStatus *brokerStatus, jk *jolokia_client.JkInfo) ArtemisError) ArtemisError {
+func (reconciler *BrokerReconcilerImpl) CheckStatusFromJolokia(jk *jolokia_client.JkInfo, checkBrokerStatus func(BrokerStatus *brokerStatus, jk *jolokia_client.JkInfo) ArtemisError) ArtemisError {
 
 	brokerStatus, artemisError := reconciler.GetAndCacheBrokerStatus(jk)
 	if artemisError != nil {
@@ -3830,7 +3831,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) CheckStatusFromJolokia(jk *jolo
 	return nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) GetAndCacheBrokerStatus(jk *jolokia_client.JkInfo) (*brokerStatus, ArtemisError) {
+func (reconciler *BrokerReconcilerImpl) GetAndCacheBrokerStatus(jk *jolokia_client.JkInfo) (*brokerStatus, ArtemisError) {
 
 	if cached, exists := reconciler.cachedBrokerStatus[jk.Ordinal]; exists {
 		switch v := cached.(type) {
@@ -3867,7 +3868,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) GetAndCacheBrokerStatus(jk *jol
 
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) resolveJolokiaEndpoints(cr *v1beta2.Broker, client rtclient.Client) {
+func (reconciler *BrokerReconcilerImpl) resolveJolokiaEndpoints(cr *v1beta2.Broker, client rtclient.Client) {
 	if reconciler.jolokiaEndpoints == nil {
 		if common.IsRestricted(cr) {
 			reconciler.jolokiaEndpoints = jolokia_client.GetMinimalJolokiaAgents(cr, client)
@@ -3886,7 +3887,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) resolveJolokiaEndpoints(cr *v1b
 	}
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) checkProjectionStatus(cr *v1beta2.Broker, client rtclient.Client, secretProjection *projection, extractStatus func(BrokerStatus *brokerStatus, FileName string) (propertiesStatus, bool)) ArtemisError {
+func (reconciler *BrokerReconcilerImpl) checkProjectionStatus(cr *v1beta2.Broker, client rtclient.Client, secretProjection *projection, extractStatus func(BrokerStatus *brokerStatus, FileName string) (propertiesStatus, bool)) ArtemisError {
 	reqLogger := ctrl.Log.WithValues("ActiveMQArtemis Name", cr.Name)
 
 	reqLogger.V(2).Info("in sync check", "projection", secretProjection)
@@ -4005,7 +4006,7 @@ func updateExtraConfigStatus(cr *v1beta2.Broker, Projection *projection) {
 		v1beta2.ExternalConfigStatus{Name: Projection.Name, ResourceVersion: Projection.ResourceVersion})
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) getSecretProjection(secretName types.NamespacedName, client rtclient.Client) (*projection, error) {
+func (reconciler *BrokerReconcilerImpl) getSecretProjection(secretName types.NamespacedName, client rtclient.Client) (*projection, error) {
 	resource := &corev1.Secret{}
 
 	// check our latest desired content
@@ -4021,7 +4022,7 @@ func (reconciler *ActiveMQArtemisReconcilerImpl) getSecretProjection(secretName 
 	return newProjectionFromByteValues(resource.ObjectMeta, resource.Data), nil
 }
 
-func (reconciler *ActiveMQArtemisReconcilerImpl) getConfigMappedJaasProperties(cr *v1beta2.Broker, client rtclient.Client) (*projection, error) {
+func (reconciler *BrokerReconcilerImpl) getConfigMappedJaasProperties(cr *v1beta2.Broker, client rtclient.Client) (*projection, error) {
 	if _, name, found := getConfigExtraMount(cr, jaasConfigSuffix); found {
 		return reconciler.getSecretProjection(types.NamespacedName{Namespace: cr.Namespace, Name: name}, client)
 	}
@@ -4137,4 +4138,774 @@ func unmarshallStatus(jsonStatus string) (brokerStatus, error) {
 func marshallApplyErrors(applyErrors []applyError) string {
 	val, _ := json.Marshal(applyErrors)
 	return string(val)
+}
+
+func (r *BrokerReconcilerImpl) validate(customResource *v1beta2.Broker, client rtclient.Client, namer common.Namers) (bool, retry bool) {
+	validationCondition := metav1.Condition{
+		Type:   v1beta2.ValidConditionType,
+		Status: metav1.ConditionTrue,
+		Reason: v1beta2.ValidConditionSuccessReason,
+	}
+
+	condition, retry := validateExtraMounts(customResource, client)
+	if condition != nil {
+		validationCondition = *condition
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse && customResource.Spec.DeploymentPlan.PodDisruptionBudget != nil {
+		condition := validatePodDisruption(customResource)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition, retry = validateNoDupKeysInBrokerProperties(customResource)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition, retry = r.validateStorage()
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition, retry = validateAcceptorPorts(customResource)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition, retry = validateSSLEnabledSecrets(customResource, client, namer)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition := common.ValidateBrokerImageVersion(customResource)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition := validateReservedLabels(customResource)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition, retry = r.validateExposeModes(customResource)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition, retry = r.validateEnvVars(customResource)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+
+	if validationCondition.Status != metav1.ConditionFalse {
+		condition, retry = r.validateRestrictedRequiredSecrets(client)
+		if condition != nil {
+			validationCondition = *condition
+		}
+	}
+	common.SetStatusConditionWithGeneration(customResource, validationCondition)
+
+	return validationCondition.Status != metav1.ConditionFalse, retry
+}
+
+func validateNoDupKeysInBrokerProperties(customResource *v1beta2.Broker) (*metav1.Condition, bool) {
+	if len(customResource.Spec.BrokerProperties) > 0 {
+		if duplicateKey := DuplicateKeyIn(customResource.Spec.BrokerProperties); duplicateKey != "" {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionFailedDuplicateBrokerPropertiesKey,
+				Message: fmt.Sprintf(".Spec.BrokerProperties has a duplicate key for %v", duplicateKey),
+			}, false
+		}
+
+	}
+	return nil, false
+}
+
+func validateReservedLabels(customResource *v1beta2.Broker) *metav1.Condition {
+	if customResource.Spec.DeploymentPlan.Labels != nil {
+		for key := range customResource.Spec.DeploymentPlan.Labels {
+			if key == selectors.LabelAppKey || key == selectors.LabelResourceKey {
+				return &metav1.Condition{
+					Type:    v1beta2.ValidConditionType,
+					Status:  metav1.ConditionFalse,
+					Reason:  v1beta2.ValidConditionFailedReservedLabelReason,
+					Message: fmt.Sprintf("'%s' is a reserved label, it is not allowed in Spec.DeploymentPlan.Labels", key),
+				}
+			}
+		}
+	}
+	for index, template := range customResource.Spec.ResourceTemplates {
+		for key := range template.Labels {
+			if key == selectors.LabelAppKey || key == selectors.LabelResourceKey {
+				return &metav1.Condition{
+					Type:    v1beta2.ValidConditionType,
+					Status:  metav1.ConditionFalse,
+					Reason:  v1beta2.ValidConditionFailedReservedLabelReason,
+					Message: fmt.Sprintf("'%s' is a reserved label, it is not allowed in Spec.DeploymentPlan.Templates[%d].Labels", key, index),
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func validateAcceptorPorts(customResource *v1beta2.Broker) (*metav1.Condition, bool) {
+	portMap := map[int32]string{}
+
+	for _, acceptor := range customResource.Spec.Acceptors {
+		if acceptor.Port > 0 {
+			if existingName, duplicate := portMap[acceptor.Port]; duplicate {
+				return &metav1.Condition{
+					Type:    v1beta2.ValidConditionType,
+					Status:  metav1.ConditionFalse,
+					Reason:  v1beta2.ValidConditionFailedDuplicateAcceptorPort,
+					Message: fmt.Sprintf(".Spec.Acceptors %q and %q contain a duplicate port %v", acceptor.Name, existingName, acceptor.Port),
+				}, false
+			}
+			portMap[acceptor.Port] = acceptor.Name
+		}
+	}
+	return nil, false
+}
+
+func (r *BrokerReconcilerImpl) validateExposeModes(customResource *v1beta2.Broker) (*metav1.Condition, bool) {
+
+	if !r.isOnOpenShift {
+		for _, acceptor := range customResource.Spec.Acceptors {
+			if acceptor.Expose && acceptor.ExposeMode != nil && *acceptor.ExposeMode == v1beta2.ExposeModes.Route {
+				return &metav1.Condition{
+					Type:    v1beta2.ValidConditionType,
+					Status:  metav1.ConditionFalse,
+					Reason:  v1beta2.ValidConditionFailedInvalidExposeMode,
+					Message: fmt.Sprintf(".Spec.Acceptors %q has invalid expose mode route, it is only supported on OpenShift", acceptor.Name),
+				}, false
+			}
+		}
+
+		for _, connector := range customResource.Spec.Connectors {
+			if connector.Expose && connector.ExposeMode != nil && *connector.ExposeMode == v1beta2.ExposeModes.Route {
+				return &metav1.Condition{
+					Type:    v1beta2.ValidConditionType,
+					Status:  metav1.ConditionFalse,
+					Reason:  v1beta2.ValidConditionFailedInvalidExposeMode,
+					Message: fmt.Sprintf(".Spec.Connectors %q has invalid expose mode route, it is only supported on OpenShift", connector.Name),
+				}, false
+			}
+		}
+
+		console := customResource.Spec.Console
+		if console.Expose && console.ExposeMode != nil && *console.ExposeMode == v1beta2.ExposeModes.Route {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionFailedInvalidExposeMode,
+				Message: ".Spec.Console has invalid expose mode route, it is only supported on OpenShift",
+			}, false
+		}
+	}
+
+	for _, acceptor := range customResource.Spec.Acceptors {
+		if acceptor.Expose && (acceptor.ExposeMode != nil && *acceptor.ExposeMode == v1beta2.ExposeModes.Ingress || !r.isOnOpenShift) &&
+			customResource.Spec.IngressDomain == "" && acceptor.IngressHost == "" {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionFailedInvalidIngressSettings,
+				Message: fmt.Sprintf(".Spec.Acceptors %q has invalid ingress settings, IngressHost unspecified and no Spec.IngressDomain default domain provided", acceptor.Name),
+			}, false
+		}
+	}
+
+	for _, connector := range customResource.Spec.Connectors {
+		if connector.Expose && (connector.ExposeMode != nil && *connector.ExposeMode == v1beta2.ExposeModes.Ingress || !r.isOnOpenShift) &&
+			customResource.Spec.IngressDomain == "" && connector.IngressHost == "" {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionFailedInvalidIngressSettings,
+				Message: fmt.Sprintf(".Spec.Connectors %q has invalid ingress settings, IngressHost unspecified and no Spec.IngressDomain default domain provided", connector.Name),
+			}, false
+		}
+	}
+
+	console := customResource.Spec.Console
+	if console.Expose && (console.ExposeMode != nil && *console.ExposeMode == v1beta2.ExposeModes.Ingress || !r.isOnOpenShift) &&
+		customResource.Spec.IngressDomain == "" && console.IngressHost == "" {
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionFailedInvalidIngressSettings,
+			Message: ".Spec.Console has invalid ingress settings, IngressHost unspecified and no Spec.IngressDomain default domain provided",
+		}, false
+	}
+
+	return nil, false
+}
+
+func (r *BrokerReconcilerImpl) validateEnvVars(customResource *v1beta2.Broker) (*metav1.Condition, bool) {
+
+	internalVarNames := map[string]string{
+		debugArgsEnvVarName:      debugArgsEnvVarName,
+		javaOptsEnvVarName:       javaOptsEnvVarName,
+		javaArgsAppendEnvVarName: javaArgsAppendEnvVarName,
+	}
+
+	invalidVars := []string{}
+
+	for _, envVar := range customResource.Spec.Env {
+		if _, ok := internalVarNames[envVar.Name]; ok {
+			if envVar.ValueFrom != nil {
+				invalidVars = append(invalidVars, envVar.Name)
+			}
+		}
+	}
+
+	if len(invalidVars) > 0 {
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionInvalidInternalVarUsage,
+			Message: fmt.Sprintf("Don't use valueFrom on env vars that the operator can mutate: %v. Instead use a different var and refernece it in its value field.", invalidVars),
+		}, false
+	}
+	return nil, false
+}
+
+func (r *BrokerReconcilerImpl) validateRestrictedRequiredSecrets(client rtclient.Client) (*metav1.Condition, bool) {
+	if common.IsRestricted(r.customResource) {
+		retry := true
+		if _, err := common.GetOperatorClientCertSecret(client); err != nil {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionMissingResourcesReason,
+				Message: fmt.Sprintf(".Spec.Restricted is true but operator failed to locate necessary operator client certificate secret, %v", err),
+			}, retry
+		}
+		if _, err := common.GetOperatorCASecret(client); err != nil {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionMissingResourcesReason,
+				Message: fmt.Sprintf(".Spec.Restricted is true but operator failed to locate necessary operator ca secret, %v", err),
+			}, retry
+		}
+		operandCertSecretName := common.GetOperandCertSecretName(r.customResource, client)
+		if _, err := common.GetNamespacedSecret(client, operandCertSecretName, r.customResource.Namespace); err != nil {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionMissingResourcesReason,
+				Message: fmt.Sprintf(".Spec.Restricted is true but operator failed to locate necessary operand cert secret, %v", err),
+			}, retry
+		}
+	}
+	return nil, false
+}
+
+func (r *BrokerReconcilerImpl) validateStorage() (*metav1.Condition, bool) {
+
+	if r.customResource.Spec.DeploymentPlan.PersistenceEnabled {
+		if r.customResource.Spec.DeploymentPlan.Storage.Size != "" {
+			_, err := resource.ParseQuantity(r.customResource.Spec.DeploymentPlan.Storage.Size)
+			if err != nil {
+				return &metav1.Condition{
+					Type:    v1beta2.ValidConditionType,
+					Status:  metav1.ConditionFalse,
+					Reason:  v1beta2.ValidConditionFailureReason,
+					Message: fmt.Sprintf(".Spec.DeploymentPlan.Storage.Size quantity string is invalid, %v", err),
+				}, false
+			}
+		}
+	}
+	return nil, false
+}
+
+func validateSSLEnabledSecrets(customResource *v1beta2.Broker, client rtclient.Client, namer common.Namers) (*metav1.Condition, bool) {
+
+	var retry = true
+	if customResource.Spec.Console.SSLEnabled {
+
+		secretName := namer.SecretsConsoleNameBuilder.Name()
+		if customResource.Spec.Console.SSLSecret != "" {
+			secretName = customResource.Spec.Console.SSLSecret
+		}
+
+		secret := corev1.Secret{}
+		found := retrieveResource(secretName, customResource.Namespace, &secret, client)
+		if !found {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionMissingResourcesReason,
+				Message: fmt.Sprintf(".Spec.Console.SSLEnabled is true but required secret %v is not found", secretName),
+			}, retry
+		}
+
+		contextMessage := ".Spec.Console.SSLEnabled is true but required"
+		for _, key := range []string{
+			"keyStorePassword",
+			"trustStorePassword",
+		} {
+			Condition := AssertSecretContainsKey(secret, key, contextMessage)
+			if Condition != nil {
+				return Condition, retry
+			}
+		}
+
+		Condition := AssertSecretContainsOneOf(secret, []string{
+			"keyStorePath",
+			"broker.ks"}, contextMessage)
+		if Condition != nil {
+			return Condition, retry
+		}
+
+		Condition = AssertSecretContainsOneOf(secret, []string{
+			"trustStorePath",
+			"client.ts"}, contextMessage)
+		if Condition != nil {
+			return Condition, retry
+		}
+	}
+	return nil, false
+}
+
+func validatePodDisruption(customResource *v1beta2.Broker) *metav1.Condition {
+	pdb := customResource.Spec.DeploymentPlan.PodDisruptionBudget
+	if pdb.Selector != nil {
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionPDBNonNilSelectorReason,
+			Message: common.PDBNonNilSelectorMessage,
+		}
+	}
+	return nil
+}
+
+func validateExtraMounts(customResource *v1beta2.Broker, client rtclient.Client) (*metav1.Condition, bool) {
+
+	instanceCounts := map[string]int{}
+	var Condition *metav1.Condition
+	var retry bool = true
+	var ContextMessage = ".Spec.DeploymentPlan.ExtraMounts.ConfigMaps,"
+	for _, cm := range customResource.Spec.DeploymentPlan.ExtraMounts.ConfigMaps {
+		configMap := corev1.ConfigMap{}
+		found := retrieveResource(cm, customResource.Namespace, &configMap, client)
+		if !found {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionMissingResourcesReason,
+				Message: fmt.Sprintf("%v missing required configMap %v", ContextMessage, cm),
+			}, retry
+		}
+		if strings.HasSuffix(cm, loggingConfigSuffix) {
+			Condition = AssertConfigMapContainsKey(configMap, LoggingConfigKey, ContextMessage)
+			instanceCounts[loggingConfigSuffix]++
+		} else if strings.HasSuffix(cm, jaasConfigSuffix) {
+			Condition = &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionFailedExtraMountReason,
+				Message: fmt.Sprintf("%v entry %v with suffix %v must be a secret", ContextMessage, cm, jaasConfigSuffix),
+			}
+			retry = false // Cr needs an update
+		}
+		if Condition != nil {
+			return Condition, retry
+		}
+	}
+
+	ContextMessage = ".Spec.DeploymentPlan.ExtraMounts.Secrets,"
+	for _, s := range customResource.Spec.DeploymentPlan.ExtraMounts.Secrets {
+		secret := corev1.Secret{}
+		found := retrieveResource(s, customResource.Namespace, &secret, client)
+		if !found {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionMissingResourcesReason,
+				Message: fmt.Sprintf("%v missing required secret %v", ContextMessage, s),
+			}, retry
+		}
+		if strings.HasSuffix(s, loggingConfigSuffix) {
+			Condition = AssertSecretContainsKey(secret, LoggingConfigKey, ContextMessage)
+			instanceCounts[loggingConfigSuffix]++
+		} else if strings.HasSuffix(s, jaasConfigSuffix) {
+			Condition = AssertSecretContainsKey(secret, JaasConfigKey, ContextMessage)
+			if Condition == nil {
+				Condition = AssertSyntaxOkOnLoginConfigData(secret.Data[JaasConfigKey], s, ContextMessage)
+			}
+			instanceCounts[jaasConfigSuffix]++
+		} else if strings.HasSuffix(s, common.BrokerPropsSuffix) {
+			Condition = AssertNoDupKeyInProperties(secret, ContextMessage)
+		}
+		if Condition != nil {
+			return Condition, retry
+		}
+	}
+	Condition = AssertInstanceCounts(instanceCounts)
+	if Condition != nil {
+		return Condition, false // CR needs update
+	}
+
+	return nil, false
+}
+
+func AssertSyntaxOkOnLoginConfigData(SecretContentForLoginConfigKey []byte, name string, contextMessage string) *metav1.Condition {
+
+	if !MatchBytesAgainsLoginConfigRegexp(SecretContentForLoginConfigKey) {
+
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionFailedExtraMountReason,
+			Message: fmt.Sprintf("%s content of login.config key in secret %v does not match supported jaas config file syntax", contextMessage, name),
+		}
+	}
+
+	return nil
+}
+
+var loginConfigSyntaxMatcher *regexp.Regexp
+
+func MatchBytesAgainsLoginConfigRegexp(buffer []byte) bool {
+	syntaxMatchRegEx := common.GetJaasConfigSyntaxMatchRegEx()
+	if syntaxMatchRegEx == "" {
+		return true
+	}
+
+	if loginConfigSyntaxMatcher == nil {
+		loginConfigSyntaxMatcher, _ = regexp.Compile(syntaxMatchRegEx)
+	}
+	return loginConfigSyntaxMatcher.Match(buffer)
+}
+
+func AssertInstanceCounts(instanceCounts map[string]int) *metav1.Condition {
+	for key, v := range instanceCounts {
+		if v > 1 {
+			return &metav1.Condition{
+				Type:    v1beta2.ValidConditionType,
+				Status:  metav1.ConditionFalse,
+				Reason:  v1beta2.ValidConditionFailedExtraMountReason,
+				Message: fmt.Sprintf("Spec.DeploymentPlan.ExtraMounts, entry with suffix %v can only be supplied once", key),
+			}
+		}
+	}
+	return nil
+}
+
+func AssertConfigMapContainsKey(configMap corev1.ConfigMap, key string, contextMessage string) *metav1.Condition {
+	if _, present := configMap.Data[key]; !present {
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionFailedExtraMountReason,
+			Message: fmt.Sprintf("%s configmap %v must have key %v", contextMessage, configMap.Name, key),
+		}
+	}
+	return nil
+}
+
+func AssertNoDupKeyInProperties(secret corev1.Secret, contextMessage string) *metav1.Condition {
+	for key, data := range secret.Data {
+		if !strings.HasPrefix(key, UncheckedPrefix) && strings.HasSuffix(key, PropertiesSuffix) {
+			if duplicateKey := DuplicateKeyInPropertiesContent(data); duplicateKey != "" {
+				return &metav1.Condition{
+					Type:    v1beta2.ValidConditionType,
+					Status:  metav1.ConditionFalse,
+					Reason:  v1beta2.ValidConditionFailedExtraMountReason,
+					Message: fmt.Sprintf("%s properties secret %v entry %v has a duplicate key for %v", contextMessage, secret.Name, key, duplicateKey),
+				}
+			}
+		}
+	}
+	return nil
+}
+
+func DuplicateKeyInPropertiesContent(keyValues []byte) string {
+	return DuplicateKeyIn(KeyValuePairs(keyValues))
+}
+
+func DuplicateKeyIn(keyValues []string) string {
+	keysMap := map[string]string{}
+
+	for _, keyAndValue := range keyValues {
+		key := extractPropertyKey(keyAndValue)
+		_, duplicate := keysMap[key]
+		if !(duplicate) {
+			keysMap[key] = key
+		} else {
+			return key
+		}
+
+	}
+
+	return ""
+}
+
+func extractPropertyKey(keyAndValue string) string {
+	for index, c := range keyAndValue {
+		if c == '=' && index > 0 && keyAndValue[index-1] != '\\' {
+			return keyAndValue[0:index]
+		}
+	}
+	return keyAndValue
+}
+
+func AssertSecretContainsKey(secret corev1.Secret, key string, contextMessage string) *metav1.Condition {
+	isCertSecret, isValid := certutil.IsSecretFromCert(&secret)
+	if isCertSecret {
+		if isValid {
+			return nil
+		}
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionInvalidCertSecretReason,
+			Message: fmt.Sprintf("%s certificate secret %s not valid, must have keys ca.crt tls.crt tls.key", contextMessage, secret.Name),
+		}
+	}
+	if _, present := secret.Data[key]; !present {
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionFailedExtraMountReason,
+			Message: fmt.Sprintf("%s secret %v must have key %v", contextMessage, secret.Name, key),
+		}
+	}
+	return nil
+}
+
+func AssertSecretContainsOneOf(secret corev1.Secret, keys []string, contextMessage string) *metav1.Condition {
+	ok, valid := certutil.IsSecretFromCert(&secret)
+	if ok {
+		if valid {
+			return nil
+		}
+		return &metav1.Condition{
+			Type:    v1beta2.ValidConditionType,
+			Status:  metav1.ConditionFalse,
+			Reason:  v1beta2.ValidConditionInvalidCertSecretReason,
+			Message: fmt.Sprintf("%s secret %s must contain keys %v", contextMessage, secret.Name, "ca.crt,tls.crt,tls.key"),
+		}
+	}
+	for _, key := range keys {
+		_, present := secret.Data[key]
+		if present {
+			return nil
+		}
+	}
+	return &metav1.Condition{
+		Type:    v1beta2.ValidConditionType,
+		Status:  metav1.ConditionFalse,
+		Reason:  v1beta2.ValidConditionFailedExtraMountReason,
+		Message: fmt.Sprintf("%s secret %v must contain one of following keys %v", contextMessage, secret.Name, keys),
+	}
+}
+
+func retrieveResource(name, namespace string, obj rtclient.Object, client rtclient.Client) bool {
+	err := client.Get(context.TODO(), types.NamespacedName{Name: name, Namespace: namespace}, obj)
+	return err == nil
+}
+
+func hasExtraMounts(cr *v1beta2.Broker) bool {
+	if cr == nil {
+		return false
+	}
+	if len(cr.Spec.DeploymentPlan.ExtraMounts.ConfigMaps) > 0 {
+		return true
+	}
+	return len(cr.Spec.DeploymentPlan.ExtraMounts.Secrets) > 0
+}
+
+func MakeNamers(customResource *v1beta2.Broker) *common.Namers {
+	newNamers := common.Namers{
+		SsGlobalName:                  "",
+		SsNameBuilder:                 namer.NamerData{},
+		SvcHeadlessNameBuilder:        namer.NamerData{},
+		SvcPingNameBuilder:            namer.NamerData{},
+		PodsNameBuilder:               namer.NamerData{},
+		SecretsCredentialsNameBuilder: namer.NamerData{},
+		SecretsConsoleNameBuilder:     namer.NamerData{},
+		SecretsNettyNameBuilder:       namer.NamerData{},
+		LabelBuilder:                  selectors.LabelerData{},
+		GLOBAL_DATA_PATH:              "/opt/" + customResource.Name + "/data",
+	}
+	newNamers.SsNameBuilder.Base(customResource.Name).Suffix("ss").Generate()
+	newNamers.SsGlobalName = customResource.Name
+	newNamers.SvcHeadlessNameBuilder.Prefix(customResource.Name).Base("hdls").Suffix("svc").Generate()
+	newNamers.SvcPingNameBuilder.Prefix(customResource.Name).Base("ping").Suffix("svc").Generate()
+	newNamers.PodsNameBuilder.Base(customResource.Name).Suffix("container").Generate()
+	newNamers.SecretsCredentialsNameBuilder.Prefix(customResource.Name).Base("credentials").Suffix("secret").Generate()
+	if customResource.Spec.Console.SSLSecret != "" {
+		newNamers.SecretsConsoleNameBuilder.SetName(customResource.Spec.Console.SSLSecret)
+	} else {
+		newNamers.SecretsConsoleNameBuilder.Prefix(customResource.Name).Base("console").Suffix("secret").Generate()
+	}
+	newNamers.SecretsNettyNameBuilder.Prefix(customResource.Name).Base("netty").Suffix("secret").Generate()
+
+	newNamers.LabelBuilder.Base(customResource.Name).Suffix("app").Generate()
+
+	return &newNamers
+}
+
+func GetDefaultLabels(cr *v1beta2.Broker) map[string]string {
+	defaultLabelData := selectors.LabelerData{}
+	defaultLabelData.Base(cr.Name).Suffix("app").Generate()
+	return defaultLabelData.Labels()
+}
+
+func conditionsModified(desiredConditions []metav1.Condition, currentConditions []metav1.Condition) bool {
+	for _, c := range desiredConditions {
+		if !common.IsConditionPresentAndEqual(currentConditions, c) {
+			return true
+		}
+	}
+	return false
+}
+
+// Controller Errors
+
+type ArtemisError interface {
+	Error() string
+	Requeue() bool
+}
+
+type artemisStatusError struct {
+	cause     error
+	transient bool
+}
+
+type jolokiaClientNotFoundError struct {
+	cause error
+}
+
+type statusOutOfSyncError struct {
+	cause string
+}
+
+type statusOutOfSyncMissingKeyError struct {
+	cause string
+}
+
+type versionMismatchError struct {
+	cause string
+}
+
+func NewArtemisStatusError(err error, transient bool) artemisStatusError {
+	return artemisStatusError{
+		err,
+		transient,
+	}
+}
+
+func (e artemisStatusError) Error() string {
+	return e.cause.Error()
+}
+
+func (e artemisStatusError) Requeue() bool {
+	return e.transient
+}
+
+func NewJolokiaClientsNotFoundError(err error) jolokiaClientNotFoundError {
+	return jolokiaClientNotFoundError{
+		err,
+	}
+}
+
+func (e jolokiaClientNotFoundError) Error() string {
+	return errors.Wrap(e.cause, "no available Jolokia Clients found").Error()
+}
+
+func (e jolokiaClientNotFoundError) Requeue() bool {
+	return true
+}
+
+func NewStatusOutOfSyncError(err error) statusOutOfSyncError {
+	return statusOutOfSyncError{err.Error()}
+}
+
+func (e statusOutOfSyncError) Error() string {
+	return e.cause
+}
+
+func (e statusOutOfSyncError) Requeue() bool {
+	return true
+}
+
+func NewStatusOutOfSyncMissingKeyError(err error) statusOutOfSyncMissingKeyError {
+	return statusOutOfSyncMissingKeyError{err.Error()}
+}
+
+func (e statusOutOfSyncMissingKeyError) Error() string {
+	return e.cause
+}
+
+func (e statusOutOfSyncMissingKeyError) Requeue() bool {
+	return true
+}
+
+type inSyncApplyError struct {
+	cause  error
+	detail map[string]string
+}
+
+const inSyncWithErrorCause = "some properties from %v resulted in error on pod %s"
+
+func NewInSyncWithError(secretProjection *projection, pod string) *inSyncApplyError {
+	return &inSyncApplyError{
+		cause:  errors.Errorf(inSyncWithErrorCause, secretProjection.Name, pod),
+		detail: map[string]string{},
+	}
+}
+
+func (e inSyncApplyError) Requeue() bool {
+	return false
+}
+
+func (e inSyncApplyError) Error() string {
+	return fmt.Sprintf("%s : reasons: %v", e.cause.Error(), e.detail)
+}
+
+func (e *inSyncApplyError) ErrorApplyDetail(container string, reason string) {
+	existing, present := e.detail[container]
+	if present {
+		e.detail[container] = fmt.Sprintf("%s, %s", existing, reason)
+	} else {
+		e.detail[container] = reason
+	}
+}
+
+func NewVersionMismatchError(err error) versionMismatchError {
+	return versionMismatchError{err.Error()}
+}
+
+func (e versionMismatchError) Error() string {
+	return e.cause
+}
+
+func (e versionMismatchError) Requeue() bool {
+	return false
 }
