+++ controllers/broker_reconciler_test.go
@@ -163,7 +163,7 @@ func TestMapComparatorForStatefulSet(t *testing.T) {
 		Comparator: compare.SimpleComparator(),
 	}
 
-	reconciler := &ActiveMQArtemisReconcilerImpl{
+	reconciler := &BrokerReconcilerImpl{
 		log:            ctrl.Log.WithName("test"),
 		customResource: nil,
 	}
@@ -202,7 +202,7 @@ func TestMapComparatorForStatefulSet(t *testing.T) {
 
 func TestComparatorMetaAndSpec(t *testing.T) {
 
-	reconciler := &ActiveMQArtemisReconcilerImpl{
+	reconciler := &BrokerReconcilerImpl{
 		log:            ctrl.Log.WithName("test"),
 		customResource: nil,
 	}
@@ -526,7 +526,7 @@ func TestNewPodTemplateSpecForCR_IncludesDebugArgs(t *testing.T) {
 		},
 	}
 
-	reconciler := &ActiveMQArtemisReconcilerImpl{
+	reconciler := &BrokerReconcilerImpl{
 		log:            ctrl.Log.WithName("test"),
 		customResource: cr,
 	}
@@ -564,8 +564,8 @@ func TestProcess_TemplateIncludesLabelsServiceAndSecret(t *testing.T) {
 				}},
 		},
 	}
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("TestProcess_TemplateIncludesLabelsServiceAndSecret"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("TestProcess_TemplateIncludesLabelsServiceAndSecret"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -641,8 +641,8 @@ func TestProcess_TemplateIncludesLabelsSecretRegexp(t *testing.T) {
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("TestProcess_TemplateIncludesLabelsServiceAndSecret"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("TestProcess_TemplateIncludesLabelsServiceAndSecret"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -693,8 +693,8 @@ func TestProcess_TemplateDuplicateKeyReplacesOk(t *testing.T) {
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("TestProcess_TemplateDuplicateKeyReplacesOk"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("TestProcess_TemplateDuplicateKeyReplacesOk"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -723,8 +723,8 @@ func Test_Respect_existing_JAVA_OPTS_properties_def(t *testing.T) {
 		Spec:       v1beta2.BrokerSpec{},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("Test_Respect_existing_JAVA_OPTS_properties_def"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("Test_Respect_existing_JAVA_OPTS_properties_def"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -800,8 +800,8 @@ func TestProcess_TemplateKeyValue(t *testing.T) {
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -889,8 +889,8 @@ func TestProcess_TemplateCustomAttributeIngress(t *testing.T) {
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -950,8 +950,8 @@ func TestProcess_TemplateCustomAttributeMisSpellingIngress(t *testing.T) {
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 	newSS, err := reconciler.ProcessStatefulSet(cr, *namer, nil)
@@ -1011,8 +1011,8 @@ func testTemplateCustomAttributeContainerSecurityContext(t *testing.T, withCRNam
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -1064,8 +1064,8 @@ func TestProcess_TemplateCustomAttributePriorityClassName(t *testing.T) {
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	namer := MakeNamers(cr)
 
@@ -1112,8 +1112,8 @@ func TestNewPodTemplateSpecForCR_AppendsDebugArgs(t *testing.T) {
 		},
 	}
 
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log.WithName("test"), isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	newSpec, err := reconciler.PodTemplateSpecForCR(cr, common.Namers{}, &appsv1.StatefulSet{}, k8sClient)
 
@@ -1139,8 +1139,8 @@ func TestNewPodTemplateSpecForCR_IncludesImagePullSecret(t *testing.T) {
 			},
 		},
 	}
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	newSpec, err := reconciler.PodTemplateSpecForCR(cr, common.Namers{}, &appsv1.StatefulSet{}, k8sClient)
 	assert.NoError(t, err)
@@ -1175,8 +1175,8 @@ func TestNewPodTemplateSpecForCR_IncludesTopologySpreadConstraints(t *testing.T)
 			},
 		},
 	}
-	outer := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	reconciler := NewActiveMQArtemisReconcilerImpl(cr, outer)
+	outer := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	reconciler := NewBrokerReconcilerImpl(cr, outer)
 
 	newSpec, err := reconciler.PodTemplateSpecForCR(cr, common.Namers{}, &appsv1.StatefulSet{}, k8sClient)
 	assert.NoError(t, err)
@@ -1203,7 +1203,7 @@ func TestNewPodTemplateSpecForCR_IncludesContainerSecurityContext(t *testing.T) 
 		},
 	}
 
-	reconciler := &ActiveMQArtemisReconcilerImpl{
+	reconciler := &BrokerReconcilerImpl{
 		log:            ctrl.Log.WithName("test"),
 		customResource: cr,
 	}
@@ -1234,7 +1234,7 @@ func TestNewPodTemplateSpecForCR_IncludesExtraVolumes(t *testing.T) {
 		},
 	}
 
-	reconciler := &ActiveMQArtemisReconcilerImpl{
+	reconciler := &BrokerReconcilerImpl{
 		log:            ctrl.Log.WithName("test"),
 		customResource: cr,
 	}
@@ -1290,7 +1290,7 @@ func TestNewPodTemplateSpecForCR_IncludesExtraVolumesWithCustomMount(t *testing.
 		},
 	}
 
-	reconciler := &ActiveMQArtemisReconcilerImpl{
+	reconciler := &BrokerReconcilerImpl{
 		log:            ctrl.Log.WithName("test"),
 		customResource: cr,
 	}
@@ -1344,7 +1344,7 @@ func TestNewPodTemplateSpecForCR_IncludesExtraVolumeClaimTemplates(t *testing.T)
 		},
 	}
 
-	reconciler := &ActiveMQArtemisReconcilerImpl{
+	reconciler := &BrokerReconcilerImpl{
 		log:            ctrl.Log.WithName("test"),
 		customResource: cr,
 	}
@@ -1790,8 +1790,8 @@ func TestEnsureOwnerReferenceAPIVersion_NoOwnerReferences(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.ensureOwnerReferenceAPIVersion(cr, existing, candidate)
 
@@ -1830,8 +1830,8 @@ func TestEnsureOwnerReferenceAPIVersion_MatchingAPIVersion(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.ensureOwnerReferenceAPIVersion(cr, existing, candidate)
 
@@ -1870,8 +1870,8 @@ func TestEnsureOwnerReferenceAPIVersion_DifferentAPIVersion(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.ensureOwnerReferenceAPIVersion(cr, existing, candidate)
 
@@ -1917,8 +1917,8 @@ func TestEnsureOwnerReferenceAPIVersion_MultipleOwnerReferences(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.ensureOwnerReferenceAPIVersion(cr, existing, candidate)
 
@@ -1959,8 +1959,8 @@ func TestEnsureOwnerReferenceAPIVersion_DifferentBrokerName(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.ensureOwnerReferenceAPIVersion(cr, existing, candidate)
 
@@ -2007,8 +2007,8 @@ func TestCompareSecret_WithAPIVersionUpdate(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.CompareSecret(deployed, requested)
 
@@ -2050,8 +2050,8 @@ func TestCompareConfigMap_WithAPIVersionUpdate(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.CompareConfigMap(deployed, requested)
 
@@ -2105,8 +2105,8 @@ func TestCompareMetaAndSpec_WithAPIVersionUpdate(t *testing.T) {
 		},
 	}
 
-	r := NewActiveMQArtemisReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
-	ri := NewActiveMQArtemisReconcilerImpl(cr, r)
+	r := NewBrokerReconciler(&NillCluster{}, ctrl.Log, isOpenshift)
+	ri := NewBrokerReconcilerImpl(cr, r)
 
 	result := ri.CompareMetaAndSpec(deployed, requested)
